+++ ui/src/plugins/notify-error/NotifyError.js
@@ -1,11 +1,12 @@
 import { Notify } from 'quasar'
 import notifyConfig from '../../shared/notify-config.js'
 
-Notify.registerType('error', { icon: 'sym_r_cancel', ...notifyConfig })
+Notify.registerType('error', notifyConfig)
 
 export default (message, caption) => {
   Notify.create({
     caption,
+    classes: 'qas-notification--error',
     message,
     type: 'error'
   })