+++ ui/src/plugins/notify-success/NotifySuccess.js
@@ -1,11 +1,12 @@
 import { Notify } from 'quasar'
 import notifyConfig from '../../shared/notify-config.js'
 
-Notify.registerType('success', { icon: 'sym_r_check_circle', ...notifyConfig })
+Notify.registerType('success', notifyConfig)
 
 export default (message, caption) => {
   Notify.create({
     caption,
+    classes: 'qas-notification--success',
     message,
     type: 'success'
   })